:( Back to not building...
BaseActivity needs to refer to Android.OS.AsyncTask.OnPostExecute (to override and call it on the base class), and it is mentioned in the docs, but doesn't actually exist.  This may be a post to the monodroid list job (will do that now).

Then the examples project has some follow on references that are not yet complete, but a ton of references and conversion completed here that I think should be good.